+++ sunvox/__init__.py
@@ -1,8 +1,8 @@
-__version__ = "0.3.1.1.9.6.1"
+__version__ = "0.3.2.2.0.0.0"
 
 SUNVOX_COPYRIGHT_NOTICE = """\
 SunVox modular synthesizer
-Copyright (c) 2008-2020, Alexander Zolotov <nightradio@gmail.com>, WarmPlace.ru
+Copyright (c) 2008-2022, Alexander Zolotov <nightradio@gmail.com>, WarmPlace.ru
 
 Ogg Vorbis 'Tremor' integer playback codec
 Copyright (c) 2002, Xiph.org Foundation