+++ sunvox/dll.py
@@ -33,7 +33,7 @@ elif DLL_BASE is not None:
     platform = sys.platform
     if platform == "linux" and os.uname()[-1] in {"armv7l", "aarch64"}:
         platform = "raspberrypi"
-    is64bit = sys.maxsize > 2 ** 32
+    is64bit = sys.maxsize > 2**32
     key = (platform, is64bit)
     rel_path = {
         ("darwin", True): "osx/lib_x86_64/sunvox.dylib",
@@ -406,6 +406,23 @@ def load_from_memory(
 
 
 @sunvox_fn(
+    _s.sv_save,
+    [
+        c_int,
+        c_char_p,
+    ],
+    c_int,
+)
+def save(
+    slot: int,
+    name: bytes,
+) -> int:
+    """
+    save project to the file.
+    """
+
+
+@sunvox_fn(
     _s.sv_play,
     [
         c_int,
@@ -462,7 +479,7 @@ def pause(
     slot: int,
 ) -> int:
     """
-    (no official docs)
+    pause the audio stream on the specified slot
     """
 
 
@@ -477,23 +494,23 @@ def resume(
     slot: int,
 ) -> int:
     """
-    (no official docs)
+    resume the audio stream on the specified slot
     """
 
 
 @sunvox_fn(
-    _s.sv_stop,
+    _s.sv_sync_resume,
     [
         c_int,
     ],
     c_int,
 )
-def stop(
+def sync_resume(
     slot: int,
 ) -> int:
     """
-    first call - stop playing;
-    second call - reset all SunVox activity and switch the engine to standby mode.
+    wait for sync (pattern effect 0x33 on any slot)
+    and resume the audio stream on the specified slot
     """
 
 
@@ -976,7 +993,12 @@ def sampler_load_from_memory(
     c_int,
 )
 def get_number_of_modules(slot: int) -> int:
-    pass
+    """
+    get the number of module slots (not the actual number of modules).
+    The slot can be empty or it can contain a module.
+    Here is the code to determine that the module slot X is not empty:
+    ( sv_get_module_flags( slot, X ) & SV_MODULE_FLAG_EXISTS ) != 0;
+    """
 
 
 @sunvox_fn(
@@ -1029,7 +1051,8 @@ def get_module_inputs(
 ) -> int:
     """
     get pointers to the int[] arrays with the input links.
-    Number of inputs = ( module_flags & MODULE.INPUTS_MASK ) >> MODULE.INPUTS_OFF
+    Number of input links = ( module_flags & MODULE.INPUTS_MASK ) >> MODULE.INPUTS_OFF
+    (this is not the actual number of connections: some links may be empty (value = -1))
     """
 
 
@@ -1047,7 +1070,9 @@ def get_module_outputs(
 ) -> int:
     """
     get pointers to the int[] arrays with the output links.
-    Number of outputs = ( module_flags & MODULE.OUTPUTS_MASK ) >> MODULE.OUTPUTS_OFF
+    Number of output links =
+    ( module_flags & MODULE.OUTPUTS_MASK ) >> MODULE.OUTPUTS_OFF
+    (this is not the actual number of connections: some links may be empty (value = -1))
     """
 
 
@@ -1266,7 +1291,12 @@ def get_module_ctl_value(
 def get_number_of_patterns(
     slot: int,
 ) -> int:
-    pass
+    """
+    get the number of pattern slots (not the actual number of patterns).
+    The slot can be empty or it can contain a pattern.
+    Here is the code to determine that the pattern slot X is not empty:
+    sv_get_pattern_lines( slot, X ) > 0;
+    """
 
 
 @sunvox_fn(
@@ -1410,6 +1440,70 @@ def get_pattern_data(
         //get the buffer with all the pattern events (notes)
       sunvox_note* n = &data[ line_number * pat_tracks + track_number ];
       ... and then do someting with note n ...
+    """
+
+
+@sunvox_fn(
+    _s.sv_set_pattern_event,
+    [
+        c_int,
+        c_int,
+        c_int,
+        c_int,
+        c_int,
+        c_int,
+        c_int,
+        c_int,
+        c_int,
+    ],
+    c_int,
+)
+def set_pattern_event(
+    slot: int,
+    pat_num: int,
+    track: int,
+    line: int,
+    nn: int,
+    vv: int,
+    mm: int,
+    ccee: int,
+    xxyy: int,
+) -> int:
+    """
+    write the pattern event to the cell at the specified line and track
+    nn,vv,mm,ccee,xxyy are the same as the fields of sunvox_note structure.
+    Only non-negative values will be written to the pattern.
+    Return value: 0 (sucess) or negative error code.
+    """
+
+
+@sunvox_fn(
+    _s.sv_get_pattern_event,
+    [
+        c_int,
+        c_int,
+        c_int,
+        c_int,
+        c_int,
+    ],
+    c_int,
+)
+def get_pattern_event(
+    slot: int,
+    pat_num: int,
+    track: int,
+    line: int,
+    column: int,
+) -> int:
+    """
+    read a pattern event at the specified line and track
+    column (field number):
+       0 - note (NN);
+       1 - velocity (VV);
+       2 - module (MM);
+       3 - controller number or effect (CCEE);
+       4 - controller value or effect parameter (XXYY);
+    Return value: value of the specified field or negative error code.
     """